requirements.txt
-e .

tokenizers
sentencepiece
shortuuid
transformers==4.51.3
accelerate
peft
bitsandbytes
pydantic
markdown2[all]
numpy
scikit-learn
gradio==4.16.0
gradio_client==0.8.1
requests
httpx==0.24.0
uvicorn
fastapi
einops==0.6.1
einops-exts==0.0.4
timm==0.6.13
ftfy
git+https://github.com/JiuhaiChen/diffusers.git
datasets
tabulate
ninja
qwen_vl_utils
huggingface_hub
wandb


pillow>=9.0.0
python-multipart
typing-extensions
# uvicorn[standard]
# httpx[http2]
psutil>=5.9.0